requirements.txt
--index-url https://download.pytorch.org/whl/cpu

blinker==1.8.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.51.0
fsspec==2024.2.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.28.1
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sympy==1.12
threadpoolctl==3.5.0
torch==2.2.0+cpu
torchsummary==1.5.1
torchvision==0.17.0+cpu
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
Werkzeug==3.0.2
